Refactor e2e gateway into test directory so it is not included in the build
