+++ packages/api/src/tests/helpers/userHelper.ts
@@ -1,7 +1,7 @@
 import type { User } from "@moj-bichard7/common/types/User"
 import type { UserGroup } from "@moj-bichard7/common/types/UserGroup"
 import { randomUUID } from "crypto"
-import End2EndPostgresGateway from "../../services/gateways/e2ePostgresGateway"
+import End2EndPostgresGateway from "../testGateways/e2ePostgresGateway"
 import { generateTestJwtTokenAndSplit } from "./jwtHelper"
 
 export const generateJwtForStaticUser = (userGroups: UserGroup[] = []): [string, User] => {
