+++ packages/api/tsconfig.build.json
@@ -1,4 +1,4 @@
 {
   "extends": "./tsconfig.json",
-  "exclude": ["dist", "node_modules", "tests", "**/*.test.*"]
+  "exclude": ["dist", "node_modules", "src/tests", "**/*.test.*"]
 }
